add image util

+++ jacksung/utils/image.py
@@ -36,8 +36,8 @@ def border(img, point1, point2, color=(0, 0, 255), border=5):
     return img
 
 
-def make_block(h, w, color=(255, 255, 255)):
-    return np.array([[color for _ in range(w)] for _ in range(h)])
+def make_block(h, w, color=(255, 255, 255), dtype=np.float32):
+    return np.array([[color for _ in range(w)] for _ in range(h)], dtype=dtype)
 
 
 def crop_png(input_path, left, top, right, bottom):
